+++ packages/noob/src/noob/node/tube.py
@@ -2,6 +2,8 @@ import warnings
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING, Any, Union
 
+from pydantic import ConfigDict
+
 from noob.edge import Slot
 from noob.exceptions import ExtraInputWarning
 from noob.node.base import Node
@@ -24,6 +26,8 @@ class TubeNode(Node):
     _tube_spec: Union["TubeSpecification", None] = None
     _runner: Union["TubeRunner", None] = None
 
+    model_config = ConfigDict(extra="allow")
+
     @property
     def tube_spec(self) -> "TubeSpecification":
         from noob.tube import TubeSpecification
@@ -36,9 +40,13 @@ class TubeNode(Node):
     def init(self, context: RunnerContext) -> None:  # type: ignore[override]
         from noob import SynchronousRunner, Tube
 
+        input_collection = context['input_collection']
+        input_params = input_collection.get_node_params({k:v for k,v in self.spec.params.items() if k != 'tube'})
+        extra_params = {k:v for k,v in self.__pydantic_extra__.items() if k not in input_params}
+
         with warnings.catch_warnings(action="ignore", category=ExtraInputWarning):
             self._tube = Tube.from_specification(
-                self.tube, input={**context["tube"].input_collection.chain}
+                self.tube, input={**input_collection.chain, **input_params, **extra_params}
             )
         self._runner = SynchronousRunner(tube=self._tube)
         self._runner.init()