+++ packages/noob/src/noob/runner/zmq/node.py
@@ -1,6 +1,7 @@
 import asyncio
 import concurrent.futures
 import contextlib
+import inspect
 import multiprocessing as mp
 import os
 import signal
@@ -10,7 +11,7 @@ from collections.abc import AsyncGenerator
 from functools import cached_property, partial
 from itertools import count
 from types import FrameType
-from typing import Any, cast
+from typing import Any, cast, Callable
 
 import zmq
 from pydantic import ValidationError
@@ -44,7 +45,7 @@ from noob.node import Node, NodeSpecification
 from noob.scheduler import Scheduler
 from noob.state import State
 from noob.store import EventStore
-from noob.types import Epoch
+from noob.types import Epoch, RunnerContext
 from noob.utils import iscoroutinefunction_partial
 
 
@@ -523,6 +524,21 @@ class NodeRunner(EventloopMixin):
             ):
                 self.scheduler.done(ep, "assets")
             self._ready_condition.notify_all()
+
+    def get_context(self) -> RunnerContext:
+        return RunnerContext(runner=self, input_collection=self.input_collection)
+
+    def inject_context(self, fn: Callable) -> Callable:
+        """Wrap function in a partial with the runner context injected, if requested"""
+        sig = inspect.signature(fn)
+        ctx_key = [
+            k for k, v in sig.parameters.items() if
+            v.annotation and v.annotation is RunnerContext
+        ]
+        if ctx_key:
+            return partial(fn, **{ctx_key[0]: self.get_context()})
+        else:
+            return fn
 
     def _init_sockets(self) -> None:
         self._init_loop()